Refactor OUT Endpoint Flow.
This reverses the calling order for OUT endpoints: The lower, hardware-specific
layers are aware of the class type they are calling out into the higher,
hardware-independent layer.

For OUT Endpoints, the majority of execution will go from the OUT Interrupt
Handler up the Software Stack. This should remove a few vtable look-ups in that
path and allow for more inlining to happen.